% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.missing.indicator.variables.r
\name{add.missing.indicator.variables}
\alias{add.missing.indicator.variables}
\title{Add missing indicator variables function}
\usage{
add.missing.indicator.variables(study_data, outcome_name = "s30d",
  features = NULL, excluded_columns = c("tc"))
}
\arguments{
\item{study_data}{The study data as a data frame. No default}

\item{outcome_name}{The name of outcome variable. Defaults to "m30d"}

\item{features}{The names of the feature variables, as a character vector. Defaults to NULL, in which case all columns except outcome_name and those potentially listed in excluded_columns are treated as features.}

\item{excluded_columns}{The names of columns that should not be treated as features. Defaults to NULL and is then ignored. Is also ignored if features is specified.}
}
\description{
This function creates missing indicator variables for each predictor variable
(feature) with missing values
}
