% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.star.caption.r
\name{add.star.caption}
\alias{add.star.caption}
\title{Add star caption}
\usage{
add.star.caption(xtable, star_caption)
}
\arguments{
\item{xtable}{The xtable object as latex, no default.}

\item{star_caption}{The caption to be added as caption*, no default.}
}
\description{
Adds caption*
}
