% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.time.between.injury.and.arrival.r
\name{add.time.between.injury.and.arrival}
\alias{add.time.between.injury.and.arrival}
\title{Add time between injury and arrival function}
\usage{
add.time.between.injury.and.arrival(study_data, data_dictionary)
}
\arguments{
\item{study_data}{The study data frame. No default.}
}
\description{
Calculates the time between injury and arrival based on doi, toi, doar and toar
}
