% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.exclusion.criteria.r
\name{apply.exclusion.criteria}
\alias{apply.exclusion.criteria}
\title{Apply exclusion criteria function}
\usage{
apply.exclusion.criteria(study_data)
}
\arguments{
\item{study_data}{The study data as a data frame. No default}
}
\description{
This function applies the study exclusion criteria and saves the results into
the parent environments results list
}
