% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.table.r
\name{coefficients.table}
\alias{coefficients.table}
\title{Make learner weight, risk, and AUROCC table function}
\usage{
\method{coefficients}{table}(prepped_sample,
  superlearner_object_path = "./superlearner.rds")
}
\arguments{
\item{prepped_sample}{The prepped_sample list. No default.}

\item{superlearner_object_path}{The path to the SuperLearner object as generated from the SuperLearner::SuperLearner() method. Default: "./superlearner.rds"}
}
\description{
This function generates a risk, learner weight and learner-AUROCC table.
}
