% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.manuscript.r
\name{compile.manuscript}
\alias{compile.manuscript}
\title{Compile manuscript function}
\usage{
compile.manuscript(manuscript_file_name_prefix, results_path = "./",
  manuscript_file_format = "rtex", compiler = "pdflatex")
}
\arguments{
\item{manuscript_file_name_prefix}{The file name prefix of the manuscript as a character vector of length 1. No default.}

\item{results_path}{The path to the results list rds object. Defaults to "./".}

\item{manuscript_file_format}{The format of the manuscript file. Defaults to "rtex".}

\item{compiler}{The compiler to use when compiling the rtex manuscript into a pdf. Defaults to pdflatex.}
}
\description{
This function compiles the Superlearner vs clinicians manuscript
}
