% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.supporting.information.r
\name{compile.supporting.information}
\alias{compile.supporting.information}
\title{Compile supporting information function}
\usage{
compile.supporting.information(file_name_prefix, results_path = "./",
  file_format = "rtex", compiler = "pdflatex")
}
\arguments{
\item{file_name_prefix}{The file name prefix of the supporting information file.}

\item{results_path}{The path to the results list rds object. Defaults to "./".}

\item{file_format}{The format of the supporting information file. Defaults to "rtex".}

\item{compiler}{The compiler to use when compiling the rtex manuscript into a pdf. Defaults to pdflatex.}
}
\description{
This function compiles supporting information
}
