% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.and.save.gridsearch.plot.r
\name{create.and.save.gridsearch.plot}
\alias{create.and.save.gridsearch.plot}
\title{Create and save gridsearch plot function}
\usage{
create.and.save.gridsearch.plot(loss_data, sample = TRUE)
}
\arguments{
\item{loss_data}{The data frame with cutpoints and loss. No default.}

\item{sample}{Logical. If TRUE only a sample of loss data is plotted. Defaults to TRUE.}
}
\description{
Plots the results of the gridsearch for optimal cutpoints
}
