% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.classification.tables.r
\name{create.classification.tables}
\alias{create.classification.tables}
\title{Create classification tables function}
\usage{
create.classification.tables(study_sample)
}
\arguments{
\item{study_sample}{The study sample object. No default.}
}
\description{
Function to create classification tables
}
