% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.mortality.plot.r
\name{create.mortality.plot}
\alias{create.mortality.plot}
\title{Create mortality plots}
\usage{
create.mortality.plot(study_sample, save_plot_data_to_results = TRUE)
}
\arguments{
\item{study_sample}{Study sample object. No default.}

\item{save_plot_data_to_results}{Logical. If TRUE the plot data is saved to results. Defaults to TRUE.}
}
\description{
Mortality plots
}
