% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.ROCR.all.r
\name{create.ROCR.all}
\alias{create.ROCR.all}
\title{Generate ROCR plot for all SL learners}
\usage{
create.ROCR.all(prepped_sample)
}
\arguments{
\item{prepped_sample}{The prepped_sample list. No default.}
}
\description{
Plots receiver operating characteristics curves of all learners included in SL.
}
