% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.ROCR.plots.r
\name{create.ROCR.plots}
\alias{create.ROCR.plots}
\title{Create ROCR plots}
\usage{
create.ROCR.plots(study_sample, ROC_or_precrec = "ROC")
}
\arguments{
\item{study_sample}{Study sample list. No default.}

\item{ROC_or_precrec}{String. To perform ROC or precision/recall analysis. Accepted values are "ROC" or "prec_rec". No default}
}
\description{
Plots reciever operating charecteristiqs curves or precision/recall curves.
}
