% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.ROCR.plots.v2.r
\name{create.ROCR.plots.v2}
\alias{create.ROCR.plots.v2}
\title{Create ROCR plots}
\usage{
create.ROCR.plots.v2(study_sample, outcome_name, device = "eps",
  split_var = "train", train_test = TRUE, ROC_or_precrec = "ROC",
  models = c("pred_con_train", "pred_cat_train", "pred_con_test",
  "pred_cat_test", "tc"),
  pretty_names = c("SuperLearner continuous prediction",
  "SuperLearner priority levels", "SuperLearner continuous prediction",
  "SuperLearner priority levels", "Clinicians priority levels"),
  subscript = FALSE)
}
\arguments{
\item{study_sample}{Study sample list. No default.}

\item{device}{Character vector with the name of the image device to use. Passed to rocr.plot (in turn, passed to save.plot). Defaults to "eps".}

\item{split_var}{The variable used to split plots. As string. Defaults to "train".}

\item{train_test}{Logical. Is the dataset splitted in train and test set? Defaults to TRUE.}

\item{ROC_or_precrec}{String. To perform ROC or precision/recall analysis. Accepted values are "ROC" or "prec_rec". No default.}

\item{models}{Model names as character vector. Defaults to c("pred_con_train", "pred_cat_train", "pred_con_test", "pred_cat_test", "tc")}

\item{pretty_names}{Names to be used in plots. As character vector. Defaults to c("SuperLearner continuous prediction", "SuperLearner priority levels", "SuperLearner continuous prediction", "SuperLearner priority levels", "Clinicians priority levels")}

\item{subscript}{Logical. If TRUE, underscores in pretty names in converted to expression. Passed to rocr.plots. Defaults to FALSE.}
}
\description{
Plots reciever operating charecteristiqs curves or precision/recall curves.
}
