% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.table.of.sample.characteristics.r
\name{create.table.of.sample.characteristics}
\alias{create.table.of.sample.characteristics}
\title{Create sample characteristics table}
\usage{
create.table.of.sample.characteristics(data_for_table, data_dictionary,
  strata = NULL, vars = NULL, exclude_vars = NULL,
  include_overall = TRUE, include_missing_column = TRUE, digits = 1,
  save = FALSE)
}
\arguments{
\item{data_for_table}{The data to use to create the table, no default.}

\item{data_dictionary}{the data dictionary, no default.}

\item{strata}{The strata variable, defaults to NULL.}

\item{vars}{The variables to include in the table, defaults to NULL, in which case it is defined as names(data_dictionary)[sapply(data_dictionary, function(x) x$incl == "Yes")].}

\item{exclude_vars}{Character vector of variable names to exclude from table, defaults to NULL.}

\item{include_overall}{Logical and used only if strata is not NULL. defaults to TRUE in which case an overall column is also included.}

\item{include_missing_column}{Logical. If TRUE a column is missing indicating the number (%) of missing values in each variable. Details to TRUE.}

\item{digits}{Integer. Number of digits to use when rounding table entries. Defaults to 1.}

\item{save}{Logical. If TRUE the table object is also saved to disk as a .tex file. Defaults to FALSE.}
}
\description{
Creates a table of sample characteristics
}
