% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.median.imputation.r
\name{do.median.imputation}
\alias{do.median.imputation}
\title{Do median imputation function}
\usage{
do.median.imputation(study_data, missing_indicator = "_missing")
}
\arguments{
\item{study_data}{The study data frame. No default.}

\item{missing_indicator}{A character vector of length 1 with the string used to idenfity variables with missing data to be imputed. Defaults to "_missing"}
}
\description{
Performs median imputation on variables with missing data
}
