% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.from.pe.and.ci.r
\name{extract.from.pe.and.ci}
\alias{extract.from.pe.and.ci}
\title{Extract from point estimates and confidence intervals list function}
\usage{
extract.from.pe.and.ci(pe_and_ci)
}
\arguments{
\item{pe_and_ci}{The point estimates and confidence intervals list}
}
\description{
This function extracts data from the point estimates and confidence intervals
list and puts them in results
}
