% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.from.table.r
\name{extract.from.table}
\alias{extract.from.table}
\title{Extract from table function}
\usage{
extract.from.table(variable, level = NULL, strata = "overall",
  format = "\%.0f", data_object, table_object)
}
\arguments{
\item{variable}{The variable on which to extract data. No default.}

\item{level}{The variable level on which to extract data. Defaults to NULL.}

\item{strata}{The name of the strata from which to extract data. Defaults to "overall".}

\item{format}{The format to use when formatting the numbers as characters. Defaults to "%.0f".}

\item{data_object}{The data object. No default.}

\item{table_object}{The table object. No default}
}
\description{
Extracts and formats data from supplied table
}
