% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.bootstrap.samples.r
\name{generate.bootstrap.samples}
\alias{generate.bootstrap.samples}
\title{Function to generate bootstrap samples}
\usage{
generate.bootstrap.samples(study_data, bs_samples)
}
\arguments{
\item{study_data}{The study data as data frame. No default}

\item{bs_samples}{Integer. The number of bootstrap samples to be generated. No default.}
}
\description{
This function generates bootstrap samples from the study_data.
}
