% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.confidence.intervals.r
\name{generate.confidence.intervals}
\alias{generate.confidence.intervals}
\title{Confidence interval function}
\usage{
generate.confidence.intervals(study_sample, model_or_pointestimate, func,
  samples, diffci_or_ci, outcome_name)
}
\arguments{
\item{study_sample}{The study sample list. No default.}

\item{model_or_pointestimate}{Character vector describing predictions or point estimates to analyse. For example, pred_cat and clinicians_predictions predictions for model.review.AUROCC or NRI+ and NRI- point estimates for model.review.reclassification. No default.}

\item{func}{Function that generates key statistic. For example, model.review.AUROCC that generates AUROCC of a given model, or model.review.reclassification that generates reclassification elements. No default.}

\item{samples}{Samples as prepared with train.predict.bssamples. No default.}

\item{diffci_or_ci}{String. Whether to return confidence interval on difference of model_or_pointestimates, or return confidence intervals on model_or_pointestimates separately or no confidence intervals. Must be one of c("diff", "ci", "none")}

\item{outcome_name}{Name of outcome variable. No default.}
}
\description{
This function generates confidence intervals around difference of two given point estimates using empirical bootstrapping.
}
