% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.confidence.intervals.v2.r
\name{generate.confidence.intervals.v2}
\alias{generate.confidence.intervals.v2}
\title{Confidence interval function}
\usage{
generate.confidence.intervals.v2(study_sample, model_names, the_func, samples,
  diffci_or_ci, outcome_name, digits = 2, ...)
}
\arguments{
\item{study_sample}{The study sample list. No default.}

\item{model_names}{Character vector. Names model names. No default.}

\item{the_func}{Function that generates key statistic. For example, model.review.AUROCC that generates AUROCC of given model(s), or model.review.reclassification that generates reclassification elements of given model(s). No default.}

\item{samples}{Samples prepared as the study_sample. List of study_sample lists. No default.}

\item{diffci_or_ci}{String. Whether to return confidence interval on difference of model_or_pointestimates, or return confidence intervals on model_or_pointestimates separately or no confidence intervals. Must be one of c("diff", "ci", "none")}

\item{outcome_name}{Name of outcome variable. No default.}

\item{digits}{Integer. Number of decimals when rounded. Defaults to 2.}
}
\description{
This function generates confidence intervals around point estimate(s) or around difference of two point estimates.
}
