% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.predictions.bssamples.r
\name{generate.predictions.bssamples}
\alias{generate.predictions.bssamples}
\title{Function to generate predictions on each sample}
\usage{
generate.predictions.bssamples(samples,
  prediction_func = "predictions.with.superlearner", parallel = FALSE,
  n_cores = NULL, log = TRUE, boot = FALSE, write_to_disk = FALSE)
}
\arguments{
\item{samples}{Bootstrap samples as prepared by prep.bssamples. No default.}

\item{prediction_func}{Name of function that generates predictions. Defaults to "predictions.with.superlearner"}

\item{parallel}{Logical. If TRUE the training will be run in parallel. Defaults to FALSE.}

\item{n_cores}{Integer. The number of cores to use for parallel computing. Defaults to NULL.}

\item{log}{Logical. Passed to predictions.with.superlearner. If TRUE progress is logged in logfile. Defaults to TRUE.}

\item{boot}{Logical. Passed to predictions.with.superlearner. Affects only what is printed to logfile. If TRUE prepped_sample is assumed to be a bootstrap sample. Defaults to FALSE.}

\item{write_to_disk}{Logical. Passed to predictions.with.superlearner. If TRUE the prediction data is saved as RDS to disk. Defaults to FALSE}
}
\description{
This makes predictions on the input samples using a prediciton function as parameter.
}
