% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.data.dictionary.r
\name{get.data.dictionary}
\alias{get.data.dictionary}
\title{Get data dictionary}
\usage{
get.data.dictionary(path = "extdata/", file_name = "data_dictionary.csv")
}
\arguments{
\item{path}{The path to the data dictionary. Defaults to "extdata/".}

\item{file_name}{The name of the data dictionary file. Defaults to "data_dictionary.csv".}
}
\description{
This function reads the data dictionary from a .csv file and returns it as a list
}
