% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch.breaks.r
\name{gridsearch.breaks}
\alias{gridsearch.breaks}
\title{Gridsearch breaks function}
\usage{
gridsearch.breaks(predictions, outcomes, grid = seq(0.01, 0.99, 0.01),
  n_breaks = 3, loss_function = "auc", sample = TRUE, parallel = FALSE,
  n_cores = NULL, save_plot = FALSE, verbose = FALSE)
}
\arguments{
\item{predictions}{A numeric vector of predicted probabilites. No default.}

\item{outcomes}{A numeric vector of 0 and 1. No default.}

\item{grid}{All values to be combnd. As numeric vector. Defaults to seq(0.01, 0.99, 0.01).}

\item{n_breaks}{Number of breaks, i.e. cutpoints. Defaults to 3.}

\item{loss_function}{The name of the loss function as a character vector. For now any performance measure included in ROCR is accepted. Defaults to "auc".}

\item{sample}{Logical. If TRUE loss is only calculated for 10% of possible breaks. Defaults to TRUE.}

\item{parallel}{Logical. If TRUE the gridsearch is done in parallel. Defaults to FALSE.}

\item{n_cores}{Integer. The number of cores to use for parallel computing. Defaults to NULL.}

\item{save_plot}{Logical. If TRUE a wireframe plot is saved to disk. Defaults to FALSE.}

\item{verbose}{Logical. If TRUE then information is printed as the gridsearch runs. Defaults to FALSE. Currently ignored.}
}
\description{
Performs a gridsearch to identify breaks to optimise a provided loss function
}
