% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.relevant.variables.r
\name{keep.relevant.variables}
\alias{keep.relevant.variables}
\title{Keep relevant variables}
\usage{
keep.relevant.variables(study_data, data_dictionary = NULL,
  variables_to_keep = NULL, variables_to_drop = NULL)
}
\arguments{
\item{study_data}{The study data as a data frame. No default}

\item{data_dictionary}{The data dictionary. Defaults to NULL. If provided, variables_to_keep will be defined as unlist(lapply(data_dictionary, function(x) if(x$c == "No") x$vn))}

\item{variables_to_keep}{Character vector of variables to keep. Defaults to NULL. Ignored if data_dictionary is not NULL.}

\item{variables_to_drop}{Character vector of variables to drop. Defaults to NULL. Currently ignored.}
}
\description{
This function takes as only input the study data data frame and keeps only
those variables that are relevant to the study
}
