% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.review.with.rocr.r
\name{model.review.with.rocr}
\alias{model.review.with.rocr}
\title{Review model with rocr package function}
\usage{
model.review.with.rocr(study_sample, outcome_name, model_names = c("pred_cat",
  "tc"), measure = "f")
}
\arguments{
\item{study_sample}{The study sample list. No default.}

\item{outcome_name}{The name of the outcome variable. No default.}

\item{model_names}{Character vector with list name of predictions. Default: c('pred_cat', 'tc')}

\item{measure}{String. Performance measure for ROCR prediction object. Defaults to "auc".}
}
\description{
This function reviews a model using the rocr package.
}
