% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.bssamples.r
\name{prep.bssamples}
\alias{prep.bssamples}
\title{Function to prepare data on bootstrap samples}
\usage{
prep.bssamples(samples)
}
\arguments{
\item{samples}{Bootstrap samples as list of data frames. No default.}
}
\description{
This function prepares every bootstrap sample for SuperLearner computations with prep.data.for.superlearner.
}
