% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{report}
\alias{report}
\title{Report function}
\usage{
report(variable, level = NULL, strata = "overall",
  data_object = study_data,
  table_object = results$raw_table_of_sample_characteristics)
}
\arguments{
\item{variable}{The name of the variable to report on. No default.}

\item{level}{The name of the level to report on. Defaults to NULL.}

\item{strata}{The name of the strata from which to extract data. Defaults to "overall".}

\item{data_object}{The data object. Defaults to study_data.}

\item{table_object}{The table object. Defaults to results$raw_table_of_sample_characteristics.}
}
\description{
Function to format results for reporting in the results section
}
