% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocr.plotter.r
\name{rocr.plot}
\alias{rocr.plot}
\title{ROCR plotter}
\usage{
rocr.plot(plot_data, device = eps, y_name = NULL, x_name = NULL,
  ylab = NULL, xlab = NULL, file_name = NULL, return_plot = FALSE,
  subscript = FALSE)
}
\arguments{
\item{plot_data}{The plot data object. No default.}

\item{device}{Character vector with the name of the image device to use. Passed to save.plot. Defaults to "eps".}

\item{y_name}{The name of the y variable. Defaults to NULL in which case the first variable is used.}

\item{x_name}{The name of the x variable. Defaults to NULL in which case the second variable is used.}

\item{ylab}{The y label. Defaults to NULL in which case the variable name is used.}

\item{xlab}{The x label. Defaults to NULL in which case the variable name is used.}

\item{file_name}{If not NULL the plot object will be saved to disk using the name provided. Defaults to NULL.}

\item{return_plot}{Logical. If TRUE the plot object is returned. Defaults to FALSE.}

\item{subscript}{Logical. If TRUE, underscores in pretty names in converted to expression. Defaults to FALSE.}
}
\description{
Plots ROCROCR
}
