% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.plot.r
\name{save.plot}
\alias{save.plot}
\title{Save plot}
\usage{
save.plot(plot_object, file_name, width = 13.2, height = width * 0.65,
  device = "eps", convert_pdf_to_eps = TRUE)
}
\arguments{
\item{plot_object}{The ggplot object. No default.}

\item{file_name}{The plot file name. No default.}

\item{width}{The width of the plot in cm. Defaults to 13.2.}

\item{height}{The height of the plot in cm. Default to 9.}

\item{device}{Character vector with the name of the image device to use. Defaults to "eps".}

\item{convert_pdf_to_eps}{Logical. If TRUE and device = "pdf" then an attempt to convert that pdf using pdftops to eps is made. Defaults to TRUE.}
}
\description{
Wrapper for ggsave that also crops and converts to eps if necessary
}
