% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.to.outcome.r
\name{set.to.outcome}
\alias{set.to.outcome}
\title{Change outcome function}
\usage{
set.to.outcome(study_data)
}
\arguments{
\item{study_data}{The study data as a data frame. No default}
}
\description{
This function changes patient outcome data if certain conditions are met. The outcome variable is set to dead if patients were dead on discharge after 30 days or if dead after 24 hours. Moreover, s30d is set to alive if coded alive and admitted to other hospital.
}
