% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.labels.r
\name{split.labels}
\alias{split.labels}
\title{Split labels}
\usage{
\method{split}{labels}(labels)
}
\arguments{
\item{labels}{String of labels separated by comma, no default}
}
\description{
Splits labels
}
