% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.dummy.variables.r
\name{to.dummy.variables}
\alias{to.dummy.variables}
\title{Transform factor variables to dummy variables}
\usage{
to.dummy.variables(prepped_data)
}
\arguments{
\item{prepped_data.}{The study_data as prepared by prep.data.for.superlearner as list. No default.}
}
\description{
This function transforms factor variables to dummy variables.
}
